Move things around, no logic changes
