+++ engine/src/main/scala/cromwell/engine/io/nio/NioFlow.scala
@@ -4,8 +4,7 @@ import akka.actor.ActorSystem
 import akka.stream.scaladsl.Flow
 import cats.effect._
 
-import scala.util.Try
-import cloud.nio.spi.{ChecksumFailure, ChecksumResult, ChecksumSkipped, ChecksumSuccess, FileHash, HashType}
+import cloud.nio.spi.{ChecksumFailure, ChecksumResult, ChecksumSkipped, ChecksumSuccess, FileHash}
 import com.typesafe.config.Config
 import common.util.IORetry
 import cromwell.core.io._
@@ -15,10 +14,7 @@ import cromwell.engine.io.RetryableRequestSupport.{isInfinitelyRetryable, isRetr
 import cromwell.engine.io.{IoAttempts, IoCommandContext, IoCommandStalenessBackpressuring}
 import cromwell.filesystems.blob.BlobPath
 import cromwell.filesystems.drs.DrsPath
-import cromwell.filesystems.gcs.GcsPath
 import cromwell.filesystems.http.HttpPath
-import cromwell.filesystems.s3.S3Path
-import cromwell.util.TryWithResource._
 import net.ceedubs.ficus.Ficus._
 import net.ceedubs.ficus.readers.ValueReader
 
@@ -134,7 +130,7 @@ class NioFlow(parallelism: Int,
 
     def readFileAndChecksum: IO[String] =
       for {
-        fileHash <- getStoredHash(command.file)
+        fileHash <- NioHashing.getStoredHash(command.file)
         uncheckedValue <- readFile
         checksumResult <- fileHash match {
           case Some(hash) => checkHash(uncheckedValue, hash)
@@ -176,32 +172,9 @@ class NioFlow(parallelism: Int,
       case nioPath => IO(nioPath.size)
     }
 
-  private def hash(hash: IoHashCommand): IO[String] =
-    // If there is no hash accessible from the file storage system,
-    // we'll read the file and generate the hash ourselves.
-    getStoredHash(hash.file)
-      .flatMap {
-        case Some(storedHash) => IO.pure(storedHash)
-        case None => generateMd5FileHashForPath(hash.file)
-      }
-      .map(_.hash)
-
-  private def getStoredHash(file: Path): IO[Option[FileHash]] =
-    file match {
-      case gcsPath: GcsPath => getFileHashForGcsPath(gcsPath).map(Option(_))
-      case blobPath: BlobPath => getFileHashForBlobPath(blobPath)
-      case drsPath: DrsPath =>
-        IO {
-          // We assume all DRS files have a stored hash; this will throw
-          // if the file does not.
-          drsPath.getFileHash
-        }.map(Option(_))
-      case s3Path: S3Path =>
-        IO {
-          Option(FileHash(HashType.S3Etag, s3Path.eTag))
-        }
-      case _ => IO.pure(None)
-    }
+  private def hash(hash: IoHashCommand): IO[String] = {
+    NioHashing.hash(hash.file)
+  }
 
   private def touch(touch: IoTouchCommand) = IO {
     touch.file.touch()
@@ -222,28 +195,6 @@ class NioFlow(parallelism: Int,
   }
 
   private def createDirectories(path: Path) = path.parent.createDirectories()
-
-  /**
-    * Lazy evaluation of a Try in a delayed IO. This avoids accidentally eagerly evaluating the Try.
-    *
-    * IMPORTANT: Use this instead of IO.fromTry to make sure the Try will be reevaluated if the
-    * IoCommand is retried.
-    */
-  private def delayedIoFromTry[A](t: => Try[A]): IO[A] = IO[A](t.get)
-
-  private def getFileHashForGcsPath(gcsPath: GcsPath): IO[FileHash] = delayedIoFromTry {
-    gcsPath.objectBlobId.map(id => FileHash(HashType.GcsCrc32c, gcsPath.cloudStorage.get(id).getCrc32c))
-  }
-
-  private def getFileHashForBlobPath(blobPath: BlobPath): IO[Option[FileHash]] = delayedIoFromTry {
-    blobPath.md5HexString.map(md5 => md5.map(FileHash(HashType.Md5, _)))
-  }
-
-  private def generateMd5FileHashForPath(path: Path): IO[FileHash] = delayedIoFromTry {
-    tryWithResource(() => path.newInputStream) { inputStream =>
-      FileHash(HashType.Md5, org.apache.commons.codec.digest.DigestUtils.md5Hex(inputStream))
-    }
-  }
 }
 
 object NioFlow {
