+++ engine/src/main/scala/cromwell/engine/io/nio/NioHashing.scala
@@ -1,0 +1,65 @@
+package cromwell.engine.io.nio
+
+import cats.effect.IO
+import cloud.nio.spi.{FileHash, HashType}
+import cromwell.core.path.Path
+import cromwell.filesystems.blob.BlobPath
+import cromwell.filesystems.drs.DrsPath
+import cromwell.filesystems.gcs.GcsPath
+import cromwell.filesystems.s3.S3Path
+import cromwell.util.TryWithResource.tryWithResource
+
+import scala.util.Try
+
+object NioHashing {
+
+  def hash(file: Path): IO[String] =
+    // If there is no hash accessible from the file storage system,
+    // we'll read the file and generate the hash ourselves.
+    getStoredHash(file)
+      .flatMap {
+        case Some(storedHash) => IO.pure(storedHash)
+        case None => generateMd5FileHashForPath(file)
+      }
+      .map(_.hash)
+
+  def getStoredHash(file: Path): IO[Option[FileHash]] = {
+    file match {
+      case gcsPath: GcsPath => getFileHashForGcsPath(gcsPath).map(Option(_))
+      case blobPath: BlobPath => getFileHashForBlobPath(blobPath)
+      case drsPath: DrsPath =>
+        IO {
+          // We assume all DRS files have a stored hash; this will throw
+          // if the file does not.
+          drsPath.getFileHash
+        }.map(Option(_))
+      case s3Path: S3Path =>
+        IO {
+          Option(FileHash(HashType.S3Etag, s3Path.eTag))
+        }
+      case _ => IO.pure(None)
+    }
+  }
+
+  private def generateMd5FileHashForPath(path: Path): IO[FileHash] = delayedIoFromTry {
+    tryWithResource(() => path.newInputStream) { inputStream =>
+      FileHash(HashType.Md5, org.apache.commons.codec.digest.DigestUtils.md5Hex(inputStream))
+    }
+  }
+
+  private def getFileHashForGcsPath(gcsPath: GcsPath): IO[FileHash] = delayedIoFromTry {
+    gcsPath.objectBlobId.map(id => FileHash(HashType.GcsCrc32c, gcsPath.cloudStorage.get(id).getCrc32c))
+  }
+
+  private def getFileHashForBlobPath(blobPath: BlobPath): IO[Option[FileHash]] = delayedIoFromTry {
+    blobPath.md5HexString.map(md5 => md5.map(FileHash(HashType.Md5, _)))
+  }
+
+  /**
+    * Lazy evaluation of a Try in a delayed IO. This avoids accidentally eagerly evaluating the Try.
+    *
+    * IMPORTANT: Use this instead of IO.fromTry to make sure the Try will be reevaluated if the
+    * IoCommand is retried.
+    */
+  private def delayedIoFromTry[A](t: => Try[A]): IO[A] = IO[A](t.get)
+}
